removing pytorch transforms
